api/requirements.txt
pillow
matplotlib
numpy
fastapi
uvicorn
tensorflow
numpy